Add double-buffering world copy

+++ src/PhoenixRTS/Teams/Teams.cpp
@@ -7,5 +7,5 @@ Phoenix::RTS::TeamMask Phoenix::RTS::Teams::AlliesOf(WorldConstRef world, uint8 
 
 Phoenix::RTS::TeamMask Phoenix::RTS::Teams::EnemiesOf(WorldConstRef world, uint8 team)
 {
-    return All ^ (1 << team);
+    return All ^ (1u << (uint64)team);
 }
